requirements/base.txt
Django>=1.11
pytz>=2017
Sphinx>=1.6
graphviz>=0.8
docutils>=0.14
jsonfield>=2.0
django-crispy-forms>=1.6
whitenoise>=3.3
djangorestframework>=3.4
